Audio cures creation improvements.
Now the project is read in order to know which are the AudioCue files to create instead of creating them on the fly. This fixes a problem related with creating files in a folder while reading it at the same time.
Now a popup appears to show the progress of creating the audio cues for the project.