Audio manager improvements (#1)
* AudioCues are now stored in PopochiuData.cfg

AudioManager is not used anymore by the plugin, and the assignation of cues in AudioManager.gd is now based on the data in PopochiuData.cfg.

* Audio cures creation improvements.

Now the project is read in order to know which are the AudioCue files to create instead of creating them on the fly. This fixes a problem related with creating files in a folder while reading it at the same time.
Now a popup appears to show the progress of creating the audio cues for the project.

* Enabled AudioCues creation when project sources change

Now each row in the Audio tab shows the path to its file.
The new implementation for the AudioCues creation allows renaming and moving audio files without breaking anything. :D :D :D