Some major updates to PopochiuRoom and Popochiu.
Now the Graphic interface and Transition layer scenes are loaded on runtime to Popochiu.
This will make easier to work on updates for the plugin.
Now PopochiuRoom inherits from YSort instead of Node2D. This allows the plugin to rely on this class to handle z-index rendering instead of using my custom-buggy implementation.
Now Props and Characters nodes in PopochiuRoom are YSort.
PopochiuSettings can now updated the Project's display width and height.
Props in a PopochiuRoom can now be nested and still be listed (and clickable) in the dock.