Added more project config to ProjectSettings.tres
Devs can now change the test size for the game window, and set the game as 2D or Pixel.
This won't take effect in already imported textures, but maybe I'll add that functionality as part of a Setup popup.
Renamed some properties in ProjectSettings.tres to make their purpose clearer.
The list of items on start in PopochiuSettings.tres updates if devs change that in the dock while the settings are open.