Save/load now handles inventory items and rooms' state.
Added descriptions to Save and Load buttons in Toolbar.
Basic room state is now loaded when loading a saved game.
Inventory items can be discarded.
When an item is discarded, the inventory hides automatically (unless it is marked as "Always visible").
Added a function to IInventory to clear the inventory.
FIX: items_on_start was not working.