==== v1.7.1 ====
PopochiuCharacter nodes can't be edited while selected inside a PopochiuRoom.
Added Warning message so players understand why PopochiuCharacter is not editable while in a PopochiuRoom.
Added two buttons to canvas editor menu for selecting a Prop or Hotspot baseline and walk to helpers.
When a PopochiuRoom is created, a NavigationPolygonInstance with a NavPoly is already in the scene so devs don't have to add those things manually to define a Walkable area.
When a PopochiuProp is created with interaction, or a PopochiuHotspot or a PopochiuRegion are created, a CollisionPolygon2D
is automatically created.
FIX: Deleting a Prop with no script will no longer delete its room folder.
FIX: Deleting a Prop with no interaction couldn't be deleted without restarting Godot.