==== v1.7.2 ====
FIX: Plugin was not working because cyclic dependencies
Comparing to PopochiuCharacter in PopochiuInspectorPlugin.gd during instantiation was breaking the install.
Same of the comparissons to PopochiuProp and PopochiuHotspot in PopochiuPlugin.gd.
To solve the above, a new script was created for making such comparissons. This file will only be loaded
after devs have restarted the plugin for the first time.