==== v1.8.1 ====
Added `link_to_item` property to PopochiuProp so the prop itself disables, enables when the item linked to it is added or discarded from the inventory.
Active inventory item deactivates with both left and right click.
Improvements in the flow of loading the room when running the project.
Created two stop methods in the AudioManager to stop a playing audio file inside a run and outside of one.
Changes in variable names and method signatures for several scripts.
Props, Hotspots and Regions inside rooms are obtained by their group instead of checking the childrens of their corresponding container nodes.