==== v1.8.3 ====
\ u )3)u / welcome @StickGrinder and thanks for this
<
Walkable areas are now listed in the room tab + button to create new ones + custom script to add behavior.
Inspector buttons created to edit walkable areas' polygon and go back when done.
PC can move from one walkable area to another.
/>

Plugin version is shown in the footer of the dock.
Added export properties to PopochiuProp to change the number of `hframes` in its Sprite and the current frame.
Updated icons, labels and styles for "Baseline" and "Walk to point" buttons in the canvas.
Updated shoes color for walkable_area icon.
Removed temporary rooms and created the two that are used in the tutorial.
Added hero image because yes.

FIX: Hovered queue was not being cleaned when leaving a room, which causes a huge crash.