Fixed implementation of Tweens
I was wrong about how tweens are now initialized. Now tweens are created before tweening properties and killed if the class variable is a valid instance and is running.
First approach on making E.run([]) to work again without yield(). Now functions that will be able to be called inside the array of instructions will have the run_ prefix, and they will return a Callable than will be triggered by E.run().