Creation popups working!
Creation popups were lacking of calling super() on their _ready() function, and _main_dock.dir was replaced by DirAccess with its static functions.
Fixed animations in TransitionLayer.tscn, which was causing a visual bug when moving between rooms.
Created a new room and a new prop (in room 101) to test the fixes in the creation popups.
Restored titles to creation popups.
Fixed a bug that was causing the plugin not to open the Room tab when editing a PopochiuRoom scene.
Updated the style of dialog options and dialog menu.