Dialog menu now has a scrollbar
Devs can define in PopochiuSettings the max number of options to show without the vertical scrollbar.
Added Dictionary vars to ICharacter and IInventory to properly load the state for ALL characters and inventory items in the game.
Added shortcut methods to PopochiuCharacter to make them walk to a Prop, Hotspot or Point in the current room.