Porting changes in v1.9.0 to Godot 4
The game now saves/loads the state of props, hotspots, regions and walkable areas in rooms.
State of characters and inventory items are now stored in memory and can be loaded properly.
Added methods to PopochiuCharacter to make easier walking to props, hotspots and regions.
Added property to PopochiuSettings to define the max number of dialog options to show in the menu.
Now the list of options in a dialog have a scroll when the limit defined in settings is exceded.
Added properties to PopochiuClickable to count the number of clicks and right clicks.