V1.9.0 (#8)
* Navigation system updated

This fixes the deprecation warning related to Navigation2D nodes. Now it uses the new Navigation2DServer to calculate the path characters will follow to move inside PopochiuWalkableAreas.

* Props' states are now stored too

This applies during the game session and when saving/loading.
Rooms states are now saved in memory when the game starts.
FIX: WalkToHelper and BaselineHelper nodes are no longer being orphaned.

* Hotspots, Walkable areas and Regions are now stored

* Dialog menu now has a scrollbar

Devs can define in PopochiuSettings the max number of options to show without the vertical scrollbar.
Added Dictionary vars to ICharacter and IInventory to properly load the state for ALL characters and inventory items in the game.
Added shortcut methods to PopochiuCharacter to make them walk to a Prop, Hotspot or Point in the current room.

* Added context menu option to set Player Character

This option appears only in PopochiuCharacter rows in Popochiu's main tab.
Added an icon to show the current player character.
Added field to PopochiuSettings to store the script_name of the default player character.

* Minor code improvements for PopochiuObjectRow

Now items in popup menu are disabled based on their ID.
FIX: PopochiuGroup was disabling non existing menu options in PopochiuObjectRow items.

* New singletons created from objects in the project

This commit attempts to close: https://github.com/mapedorr/popochiu/issues/19
C, I, and D singletons are now loaded from files in res://popochiu/Autoloads in order to populate them with variables to make easier to access the characters, inventory items and dialog in the project.
Updated Character and Inventory item script templates so accessing the state property reads the corresponding State script, which allows for autocompletion.

* Added R singleton

The rooms in the project are added to the custom R script.

* Added property to IRoom to avoid creating orphan nodes

FIX: An error was appearing when moving between rooms. This was because of removing the map_rid in the walkable areas.

* Created v1.9.0 realese notes

* Creating scripts with state no longer shows an error

* Singletons are updated when adding main objects to core